+++ Poliedro.Eds.Application/Shopping/Commands/CreateShopping/ShoppingProductRequestDto.cs
@@ -2,8 +2,8 @@ namespace Poliedro.Eds.Application.Shopping.Commands.CreateShopping;
 
 public record ShoppingProductRequestDto(
     int IdProduct,
-    decimal Quantity,
-    decimal PurchasePrice,
-    decimal SellPrice,
+    double Quantity,
+    double PurchasePrice,
+    double SellPrice,
     int IdCompartment
 );
